+++ src/frontend/Orion/Administrator/ListsConfigManagement/ListItemsPanel/index.js
@@ -1,1 +1,4 @@
-export { default } from './ListItemsPanel'
+import ListItemsPanelListHeader from './ListItemsPanelListHeader'
+import ListItemsPanelListItem from './ListItemsPanelListItem'
+
+export { ListItemsPanelListHeader, ListItemsPanelListItem }