+++ src/frontend/components/StructuralListPanels/utils/index.js
@@ -1,0 +1,3 @@
+import { getNewSearchParams, pushNewSearchParams } from './newSearchParamsUtils'
+
+export { getNewSearchParams, pushNewSearchParams }